+++ tests/src/test/java/io/streamnative/pulsar/handlers/kop/metrics/KafkaProducerStatsTest.java
@@ -11,12 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.pulsar.handlers.kop;
+package io.streamnative.pulsar.handlers.kop.metrics;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNotEquals;
 
 import com.google.common.collect.Sets;
+import io.streamnative.pulsar.handlers.kop.KopProtocolHandlerTestBase;
 import java.util.Properties;
 import lombok.Cleanup;
 import lombok.extern.slf4j.Slf4j;
