+++ tests/src/test/java/io/streamnative/pulsar/handlers/kop/metrics/MetricsProviderWithDisableGroupLevelConsumerMetricsTest.java
@@ -11,7 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.pulsar.handlers.kop;
+package io.streamnative.pulsar.handlers.kop.metrics;
+
+import io.streamnative.pulsar.handlers.kop.metrics.MetricsProviderTest;
 
 public class MetricsProviderWithDisableGroupLevelConsumerMetricsTest extends MetricsProviderTest {
     public MetricsProviderWithDisableGroupLevelConsumerMetricsTest() {
