+++ tests/src/test/java/io/streamnative/pulsar/handlers/kop/producer/PreciselyMessagePublishThrottlingTest.java
@@ -11,9 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.pulsar.handlers.kop;
-
-import io.streamnative.pulsar.handlers.kop.producer.MessagePublishThrottlingTest;
+package io.streamnative.pulsar.handlers.kop.producer;
 
 /**
  * Test KoP precisely messages publish throttling.
